requirements.txt
rethinkdb>=2.0
redis>=2.10
pyaml>=15.5
